test(pipnn): keep regression tooling local
Remove the submitted DiskANN microbenchmark target and co-locate numerical tests with their implementation files.
