+++ .github/workflows/ci.yml
@@ -71,7 +71,6 @@ jobs:
       - test-workspace
       - test-workspace-features
       - coverage
-      - iai-callgrind
       - vectorset-clippy
       - vectorset-fmt
       - vectorset-build
@@ -95,32 +94,6 @@ jobs:
       # - docs
     steps:
       - run: exit 0
-
-  iai-callgrind:
-    needs: basics
-    name: IAI-Callgrind microbenchmarks
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Install Rust
-        run: rustup show
-
-      - name: Install Valgrind
-        run: |
-          sudo apt-get update -qq
-          sudo apt-get install -y valgrind
-
-      - name: Install iai-callgrind-runner
-        run: cargo install iai-callgrind-runner --version 0.14.2 --locked
-
-      - uses: Swatinem/rust-cache@v2
-
-      - name: Run shared DiskANN IAI target
-        run: |
-          cargo bench --locked -p diskann \
-            --bench bench_main_iai \
-            --features pipnn,testing
 
   fmt:
     name: format check
