+++ diskann/Cargo.toml
@@ -32,7 +32,6 @@ diskann-wide = { workspace = true }
 dashmap = { workspace = true, optional = true }
 
 [dev-dependencies]
-iai-callgrind.workspace = true
 futures-util = { workspace = true, default-features = false  }
 pin-project.workspace = true
 rand.workspace = true
@@ -41,11 +40,6 @@ serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 tokio = { workspace = true, features = ["macros", "sync"] }
 dashmap = { workspace = true }
-
-[[bench]]
-name = "bench_main_iai"
-harness = false
-required-features = ["pipnn", "testing"]
 
 # Some 'cfg's in the source tree will be flagged by `cargo clippy -j 2 --workspace --no-deps --all-targets -- -D warnings`
 [lints.rust]
