--- diskann/benches/benchmarks_iai/mod.rs
@@ -1,6 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * Licensed under the MIT license.
- */
-
-pub(crate) mod pipnn_kernels;
